matquiz_python_ai_service/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
distro==1.9.0
google-auth==2.48.0
google-genai==1.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyasyncore==1.0.5
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python3-pika==0.9.14
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1

aio-pika